+++ minty-support/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log
 
-### 1.0.3
+## 1.1.0
+
+- Fix blue colors for non-.minty files.
+- Add openMintyDirectory command, which opens the local Minty directory, if any. Found at $(MINTY_PATH).
+- Update document link provider to check files within $(MINTY_PATH)/Data.
+- Add syntax highlighting for macros when used.
+
+## 1.0.4
+
+- Fix README to have correct keyboard shortcuts.
+- Rename color theme to "Minty."
+
+## 1.0.3
 
 - Adjust blue colors to be lighter.